requirements.txt
openai==0.27.11
flask==2.3.3
flask-cors==4.0.0
langchain==0.1.16
langchain-core==0.1.35
langchain-community==0.0.21
langchain-openai==0.0.8
faiss-cpu==1.11.1
python-dotenv==1.0.1
gunicorn==23.0.0
